requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-default-group==1.2.4
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.117.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
idna==3.10
multidict==6.6.4
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlite-fts4==1.0.3
sqlite-utils==3.38
starlette==0.48.0
tabulate==0.9.0
twilio==9.8.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.20.1
